Refactor bot into package with config and strategy modules
